requirements.txt
defusedxml==0.4.1
Django==1.10.1
django-allauth==0.27.0
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.9.0
django-dynamic-preferences==0.8.2
django-mptt==0.8.6
django-simple-menu==1.2.1
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
easy-thumbnails==2.3
mysqlclient==1.3.7
oauthlib==2.0.0
persisting-theory==0.2.1
Pillow==3.3.1
pkg-resources==0.0.0
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
